NPH Updates
I added a Contact form but am having trouble styling it.
I think it's because the form has a display of "block". I don't see how to change this. I copied the code from W3.
I will try and construct my own contact form and use the display:inline-block and see if that resolves my issues.